requirements.txt
streamlit==1.49.1
numpy>=1.21, <2.0
pandas==2.3.2
scipy==1.16.2
yfinance==0.2.65
dataframe_image==0.2.7
finnhub-python==2.4.24
matplotlib==3.10.6
plotly==6.3.0
python-telegram-bot==22.3
PyYAML>=5.3
streamlit-calendar==1.4.0
simfin>=1.0
psycopg2-binary==2.9.10
etf_scraper==0.1.2
playwright==1.55.0
plotly==6.3.0
sqlalchemy==2.0.43
rpy2>=3.4.0
sqlalchemy>-1.3
etf_scraper
fredapi>=0.5
holidays